pyproject.toml
[project]
name = "trp-audit-challenge"
version = "0.1.0"
description = "Automaton Auditor: A production-grade LangGraph swarm for automated code governance."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-core>=0.3.0",
    "langsmith>=0.1.0",
    "pydantic>=2.0.0",
    "gitpython>=3.1.0",
    "docling>=2.0.0",
    "pymupdf>=1.24.0",
    "python-dotenv>=1.0.0",
    "typing-extensions>=4.0.0",
    "langchain-google-genai>=4.2.1",
    "google-generativeai>=0.8.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
